end2end-benefits/benefits-claims/stream_app/app/requirements.txt
streamlit
elasticsearch
psycopg2-binary==2.9.9
python-dotenv
openai
sentence_transformers==2.7.0
numpy
pandas
sqlalchemy

--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.3.1+cpu



